june-2020/demo-website/requirements.txt
Django==2.2.10
django-admin-tools==0.8.1
django-appconf==1.0.3
django-classy-tags==0.9.0
django-compressor==2.3
django-cookie-law==2.0.2
django-extensions==2.2.5
django-mathfilters==1.0.0
django-sass-processor==0.7.4
django-simple-cookie-consent==0.1.1
djangorestframework==3.10.3
-e git+https://github.com/dprog-philippe-docourt/django-qr-code.git@d8811b7e05b1beff466c7800007b603c3c1eae7f#egg=django_qr_code
requests==2.23.0
PyJWT==1.7.1
python-decouple==3.3
Pillow==6.2.2
psycopg2-binary==2.8.5
libsass==0.19.4
